Don't compress collect log file.
Previously, we uesed the logging module's encoding=bz2_codec support to
compress the incoming tweets. However, pypy's logging module does not
support this encoding as of this patch. To work around it, we collect
the incming tweets uncompressed and then compress them before
processing.

This is the only reasonable workaround. The logging module forces a
newline after each call to emit(), so we can't encode manually each
tweet using the bz2 module. If we try to do that, then the compressed
bzip2 file is corrupted due to extra newlines. Furthermore, we can't
remove these newlines since the encoded stream can have bytes with the
value of a newline having nothing to do with the formatted text.

The noweats system is working with pypy as of the patch. It is 2x faster
than the original CPython version.